requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0

# Networking and IoT
paho-mqtt==1.6.1
aiocoap==0.4.5
requests==2.31.0
asyncio==3.4.3

# Security
cryptography==41.0.3
pyOpenSSL==23.2.0
PyJWT==2.8.0
paramiko==3.3.1

# Cloud Services
boto3==1.28.25
azure-identity==1.14.0
google-cloud-storage==2.10.0

# Data Processing
scipy==1.11.2
imbalanced-learn==0.11.0  # For SMOTE

# Monitoring and Logging
prometheus-client==0.17.1
elasticsearch==8.9.0
python-json-logger==2.0.7

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# ML Explainability
shap==0.42.1
lime==0.2.0.1

# Network Simulation
scapy==2.5.0
mininet==2.3.0  # Note: Linux only, use WSL on Windows

# Policy Engine
opa-python-client==1.3.0

# Utilities
click==8.1.7
tqdm==4.66.1
colorama==0.4.6
python-dateutil==2.8.2
